rename file to scenario_utils
