+++ docs/files/api/input.rst
@@ -8,7 +8,7 @@ Input data
     zen_garden.input.element_data_loader
     zen_garden.input.dataset_path_resolver
     zen_garden.input.scenario_dict
-    zen_garden.input.scenario_expansion
+    zen_garden.input.scenario_utils
     zen_garden.input.input_data_checks
     zen_garden.input.network_topology
     zen_garden.input.unit_converter
