+++ zen_garden/utils/string_utils.py
@@ -3,7 +3,7 @@ import os
 import shutil
 from pathlib import Path
 
-from zen_garden.input.scenario_expansion import ScenarioUtils
+from zen_garden.input.scenario_utils import ScenarioUtils
 
 logger = logging.getLogger(__name__)
